requirements.txt
asgiref==3.2.3
Django==3.0
django-ckeditor==5.8.0
django-cors-headers==3.2.0
django-filter==2.2.0
django-js-asset==1.2.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
Faker==3.0.0
mixins==0.1.1
Pillow==6.2.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
